+++ src/Installer/Modules/DevLogging.php
@@ -29,17 +29,12 @@ class DevLogging extends Module
             <label for="clear" class="col-xs-2 control-label">Logs löschen</label>
             <div class="col-xs-6">
             ';
-        
-        if (count(scandir($this->config->platform_root . '/jtlconnector/logs')) > 3){
-            $html .=' <button formmethod="post" formaction="/jtlconnector/install/loggingConfig.php" name="clear" class="btn btn-default btn-sm btn-block">Clear</button>';
-        }else{
-            $html .= '
-            <div data-toggle="tooltip" data-placement="top" title="Es wurden keine Logs gefunden!">
-                <button disabled class="btn btn-default btn-sm btn-block">Clear</button>
-            </div>
-            
-            ';
+    
+        $disabledButtons = "";
+        if (count(scandir($this->config->platform_root . '/jtlconnector/logs')) < 3) {
+            $disabledButtons = "disabled";
         }
+        $html .= '<div id="clearLogsButton" '. $disabledButtons .' class="btn btn-default btn-sm btn-block">Clear</div>';
         
         $html .='
             </div>
@@ -48,19 +43,34 @@ class DevLogging extends Module
             <label for="download" class="col-xs-2 control-label">Logs herunterladen</label>
             <div class="col-xs-6">';
         
-        if (count(scandir($this->config->platform_root . '/jtlconnector/logs')) > 3){
-            $html .='<button formmethod="post" formaction="/jtlconnector/install/loggingConfig.php" name="download" class="btn btn-default btn-sm btn-block">Download</button>';
-        }else{
-            $html .= '
-            <div data-toggle="tooltip" data-placement="top" title="Es wurden keine Logs gefunden!">
-                <button disabled class="btn btn-default btn-sm btn-block">Download</button>
-            </div>
-            
-            ';
-        }
+        $html .= '<div id="downloadLogsButton" '. $disabledButtons .' class="btn btn-default btn-sm btn-block">Download</div>';
         
         $html .= '</div>
         </div>
+            <script>
+                document.getElementById("downloadLogsButton").addEventListener("click", downloadLogs)
+                document.getElementById("clearLogsButton").addEventListener("click", clearLogs)
+                
+                function downloadLogs() {
+                    $.ajax({
+                      type: "POST",
+                      url: " '. substr($this->shopConfig["shop"]["folder"],0, -1) .'/jtlconnector/install/loggingConfig.php",
+                      data: {download: ""},
+                    }).done(() => {
+                        window.location.href = "'. substr($this->shopConfig["shop"]["folder"],0, -1) .'/jtlconnector/install/logs.zip";
+                    });
+                }
+                
+                function clearLogs() {
+                    $.ajax({
+                      type: "POST",
+                      url: " '. substr($this->shopConfig["shop"]["folder"],0, -1) .'/jtlconnector/install/loggingConfig.php",
+                      data: {clear: ""},
+                    }).done(() => {
+                        location.reload();
+                    });
+                }
+            </script>
         </div>';
         
         return $html;